requirements.txt
Django==1.9.8
GroupyAPI==0.6.6
Pillow==2.5.3
cleverbot==1.1.0
dj-database-url==0.4.1
future==0.16.0
gunicorn==19.6.0
psycopg2==2.6.2
python-firebase==1.2
requests==2.11.1
whitenoise==3.2
